Refactor Jira API integration: move fetchMe and fetchProjects functions to a new jira.ts file, streamline imports, and enhance error handling.
